use tigher 'int' types for ctbAddr like variables

+++ libde265/pps.h
@@ -146,10 +146,10 @@ public:
   int Log2MinCuChromaQpOffsetSize;
   int Log2MaxTransformSkipSize;
 
-  int colWidth [ DE265_MAX_TILE_COLUMNS ];
-  int rowHeight[ DE265_MAX_TILE_ROWS ];
-  int colBd    [ DE265_MAX_TILE_COLUMNS+1 ];
-  int rowBd    [ DE265_MAX_TILE_ROWS+1 ];
+  uint16_t colWidth [ DE265_MAX_TILE_COLUMNS ];
+  uint16_t rowHeight[ DE265_MAX_TILE_ROWS ];
+  uint16_t colBd    [ DE265_MAX_TILE_COLUMNS+1 ];
+  uint16_t rowBd    [ DE265_MAX_TILE_ROWS+1 ];
 
   std::vector<uint32_t> CtbAddrRStoTS; // #CTBs
   std::vector<uint32_t> CtbAddrTStoRS; // #CTBs
